+++ packages/d6-system-2e-core-content/d6-system-2e-core-content.mjs
@@ -16,7 +16,7 @@ Hooks.once("ready", () => {
     mechanicIds: [],
     recommendedPrimaryProfile: "second-edition",
     rulesFamily: "d6-system-second-edition",
-    version: "0.1.0-alpha.30"
+    version: "0.1.0-alpha.31"
   });
 });
 //# sourceMappingURL=d6-system-2e-core-content.mjs.map